@@ -6,29 +6,33 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
+  cached_network_image: ^3.2.0
+  cupertino_icons: ^1.0.4
   flutter:
     sdk: flutter
-  cached_network_image: ^2.3.3
-  cupertino_icons: ^0.1.3
-  http: ^0.12.2
-  json_annotation: ^3.1.0
-  progressive_image: ^1.0.1
+  freezed: ^2.0.3
+  http: ^0.13.4
+  json_annotation: ^4.5.0
+  openapi_base: ^1.0.2+2
+  openapi_generator_annotations: ^3.3.0
+  progressive_image: ^2.0.0
 
 dev_dependencies:
+  build_runner: ^2.1.10
+  flutter_launcher_icons: ^0.9.2
   flutter_test:
     sdk: flutter
-  build_runner: ^1.10.1
-  flutter_launcher_icons: ^0.8.1
-  json_serializable: ^3.5.0
+  json_serializable: ^6.2.0
+  #openapi_freezed_dio_builder: any
+  openapi_generator: ^3.3.0+1
 
 flutter_icons:
   android: true
   ios: true
   image_path: "assets/app_alt.png"
-
 
 flutter:
   uses-material-design: true
