@@ -21,10 +21,10 @@ class Creators {
     this.results,
   });
 
-  int count;
-  String next;
+  int? count;
+  String? next;
   dynamic previous;
-  List<Result> results;
+  List<Result>? results;
 
   factory Creators.fromJson(Map<String, dynamic> json) =>
       _$CreatorsFromJson(json);
@@ -45,14 +45,14 @@ class Result {
     this.games,
   });
 
-  int id;
-  String name;
-  String slug;
-  String image;
-  String imageBackground;
-  int gamesCount;
-  List<Game> positions;
-  List<Game> games;
+  int? id;
+  String? name;
+  String? slug;
+  String? image;
+  String? imageBackground;
+  int? gamesCount;
+  List<Game>? positions;
+  List<Game>? games;
 
   factory Result.fromJson(Map<String, dynamic> json) => _$ResultFromJson(json);
 
@@ -68,10 +68,10 @@ class Game {
     this.added,
   });
 
-  int id;
-  String slug;
-  String name;
-  int added;
+  int? id;
+  String? slug;
+  String? name;
+  int? added;
 
   factory Game.fromJson(Map<String, dynamic> json) => _$GameFromJson(json);
 
