@@ -21,10 +21,10 @@ class Stores {
     this.results,
   });
 
-  int count;
+  int? count;
   dynamic next;
   dynamic previous;
-  List<Result> results;
+  List<Result>? results;
 
   factory Stores.fromJson(Map<String, dynamic> json) => _$StoresFromJson(json);
 
@@ -43,13 +43,13 @@ class Result {
     this.games,
   });
 
-  int id;
-  String name;
-  String domain;
-  String slug;
-  int gamesCount;
-  String imageBackground;
-  List<Game> games;
+  int? id;
+  String? name;
+  String? domain;
+  String? slug;
+  int? gamesCount;
+  String? imageBackground;
+  List<Game>? games;
 
   factory Result.fromJson(Map<String, dynamic> json) => _$ResultFromJson(json);
 
@@ -65,10 +65,10 @@ class Game {
     this.added,
   });
 
-  int id;
-  String slug;
-  String name;
-  int added;
+  int? id;
+  String? slug;
+  String? name;
+  int? added;
 
   factory Game.fromJson(Map<String, dynamic> json) => _$GameFromJson(json);
 
