@@ -6,43 +6,36 @@ part of 'creator.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Creators _$CreatorsFromJson(Map<String, dynamic> json) {
-  return Creators(
-    count: json['count'] as int,
-    next: json['next'] as String,
-    previous: json['previous'],
-    results: (json['results'] as List)
-        ?.map((e) =>
-            e == null ? null : Result.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-  );
-}
+Creators _$CreatorsFromJson(Map<String, dynamic> json) => Creators(
+      count: json['count'] as int?,
+      next: json['next'] as String?,
+      previous: json['previous'],
+      results: (json['results'] as List<dynamic>?)
+          ?.map((e) => Result.fromJson(e as Map<String, dynamic>))
+          .toList(),
+    );
 
 Map<String, dynamic> _$CreatorsToJson(Creators instance) => <String, dynamic>{
       'count': instance.count,
       'next': instance.next,
       'previous': instance.previous,
-      'results': instance.results?.map((e) => e?.toJson())?.toList(),
+      'results': instance.results?.map((e) => e.toJson()).toList(),
     };
 
-Result _$ResultFromJson(Map<String, dynamic> json) {
-  return Result(
-    id: json['id'] as int,
-    name: json['name'] as String,
-    slug: json['slug'] as String,
-    image: json['image'] as String,
-    imageBackground: json['imageBackground'] as String,
-    gamesCount: json['gamesCount'] as int,
-    positions: (json['positions'] as List)
-        ?.map(
-            (e) => e == null ? null : Game.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-    games: (json['games'] as List)
-        ?.map(
-            (e) => e == null ? null : Game.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-  );
-}
+Result _$ResultFromJson(Map<String, dynamic> json) => Result(
+      id: json['id'] as int?,
+      name: json['name'] as String?,
+      slug: json['slug'] as String?,
+      image: json['image'] as String?,
+      imageBackground: json['imageBackground'] as String?,
+      gamesCount: json['gamesCount'] as int?,
+      positions: (json['positions'] as List<dynamic>?)
+          ?.map((e) => Game.fromJson(e as Map<String, dynamic>))
+          .toList(),
+      games: (json['games'] as List<dynamic>?)
+          ?.map((e) => Game.fromJson(e as Map<String, dynamic>))
+          .toList(),
+    );
 
 Map<String, dynamic> _$ResultToJson(Result instance) => <String, dynamic>{
       'id': instance.id,
@@ -51,18 +44,16 @@ Map<String, dynamic> _$ResultToJson(Result instance) => <String, dynamic>{
       'image': instance.image,
       'imageBackground': instance.imageBackground,
       'gamesCount': instance.gamesCount,
-      'positions': instance.positions?.map((e) => e?.toJson())?.toList(),
-      'games': instance.games?.map((e) => e?.toJson())?.toList(),
+      'positions': instance.positions?.map((e) => e.toJson()).toList(),
+      'games': instance.games?.map((e) => e.toJson()).toList(),
     };
 
-Game _$GameFromJson(Map<String, dynamic> json) {
-  return Game(
-    id: json['id'] as int,
-    slug: json['slug'] as String,
-    name: json['name'] as String,
-    added: json['added'] as int,
-  );
-}
+Game _$GameFromJson(Map<String, dynamic> json) => Game(
+      id: json['id'] as int?,
+      slug: json['slug'] as String?,
+      name: json['name'] as String?,
+      added: json['added'] as int?,
+    );
 
 Map<String, dynamic> _$GameToJson(Game instance) => <String, dynamic>{
       'id': instance.id,
