@@ -6,41 +6,35 @@ part of 'platform.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Platforms _$PlatformsFromJson(Map<String, dynamic> json) {
-  return Platforms(
-    count: json['count'] as int,
-    next: json['next'] as String,
-    previous: json['previous'],
-    results: (json['results'] as List)
-        ?.map((e) =>
-            e == null ? null : Result.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-  );
-}
+Platforms _$PlatformsFromJson(Map<String, dynamic> json) => Platforms(
+      count: json['count'] as int?,
+      next: json['next'] as String?,
+      previous: json['previous'],
+      results: (json['results'] as List<dynamic>?)
+          ?.map((e) => Result.fromJson(e as Map<String, dynamic>))
+          .toList(),
+    );
 
 Map<String, dynamic> _$PlatformsToJson(Platforms instance) => <String, dynamic>{
       'count': instance.count,
       'next': instance.next,
       'previous': instance.previous,
-      'results': instance.results?.map((e) => e?.toJson())?.toList(),
+      'results': instance.results?.map((e) => e.toJson()).toList(),
     };
 
-Result _$ResultFromJson(Map<String, dynamic> json) {
-  return Result(
-    id: json['id'] as int,
-    name: json['name'] as String,
-    slug: json['slug'] as String,
-    gamesCount: json['gamesCount'] as int,
-    imageBackground: json['imageBackground'] as String,
-    image: json['image'],
-    yearStart: json['yearStart'] as int,
-    yearEnd: json['yearEnd'],
-    games: (json['games'] as List)
-        ?.map(
-            (e) => e == null ? null : Game.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-  );
-}
+Result _$ResultFromJson(Map<String, dynamic> json) => Result(
+      id: json['id'] as int?,
+      name: json['name'] as String?,
+      slug: json['slug'] as String?,
+      gamesCount: json['gamesCount'] as int?,
+      imageBackground: json['imageBackground'] as String?,
+      image: json['image'],
+      yearStart: json['yearStart'] as int?,
+      yearEnd: json['yearEnd'],
+      games: (json['games'] as List<dynamic>?)
+          ?.map((e) => Game.fromJson(e as Map<String, dynamic>))
+          .toList(),
+    );
 
 Map<String, dynamic> _$ResultToJson(Result instance) => <String, dynamic>{
       'id': instance.id,
@@ -51,17 +45,15 @@ Map<String, dynamic> _$ResultToJson(Result instance) => <String, dynamic>{
       'image': instance.image,
       'yearStart': instance.yearStart,
       'yearEnd': instance.yearEnd,
-      'games': instance.games?.map((e) => e?.toJson())?.toList(),
+      'games': instance.games?.map((e) => e.toJson()).toList(),
     };
 
-Game _$GameFromJson(Map<String, dynamic> json) {
-  return Game(
-    id: json['id'] as int,
-    slug: json['slug'] as String,
-    name: json['name'] as String,
-    added: json['added'] as int,
-  );
-}
+Game _$GameFromJson(Map<String, dynamic> json) => Game(
+      id: json['id'] as int?,
+      slug: json['slug'] as String?,
+      name: json['name'] as String?,
+      added: json['added'] as int?,
+    );
 
 Map<String, dynamic> _$GameToJson(Game instance) => <String, dynamic>{
       'id': instance.id,
